Modularized Plot logic Chart (orchestrator) ChartBar + ChartDoc (placeholder)  #32
